+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to
 
 - ♿(frontend) use aria-haspopup menu on DropButton triggers #2126
 - ♿️(frontend) add contextual browser tab titles for docs routes #2120
+- ♿️(frontend) fix empty heading before section titles in HTML export #2125
 
 ## [v4.8.4] - 2026-03-25